requirements.txt
fastapi
uvicorn
pydantic

# LangChain
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-text-splitters

# Embeddings & Reranker
sentence-transformers
transformers
torch

# Vector Store
pinecone

# Document Loading
pypdf

# Utilities
numpy
requests
python-multipart
python-dotenv